chore: split docs content resolver into multiple files
